When find returns multiple archive results, location should be the archive.
This is not ideal (user has to go find the actual file ), but it's about
the best we can do without major compile-mode surgery. There are no hooks
for find-file and no archive file handler, and it's overkill to write one
for this purpose.

Visiting files in archive still works just fine when there's only one result,
which IMHO is the major use case.